% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bezier.r
\name{bezier}
\alias{bezier}
\title{Interpolate using Bezier curve}
\usage{
bezier(xyt, t.slice)
}
\arguments{
\item{xyt}{a 4x3 dataframe containing the coordinates and times for the four control points. Each row of the dataframe should be arranged as x, y, t.}

\item{t.slice}{a single time (POSIX or numeric), or list of times, to be interpolated for. The times must lie between those of the middle two control points (i.e., the times associated with rows 2 & 3 in the \code{xyt} dataframe).}
}
\value{
The function returns a dataframe (with \code{nrow = length{t.slice}}) corresponding to the interpolated locations.
}
\description{
Perform path interpolation using a Bezier curve interpolation method.
}
\details{
Bezier curves can be useful for interpolating movement when the object exhibits curvi-linear movement shapes (Long 2015). For example, the movements of marine mammals often exhibit this property (Tremblay \emph{et al.} 2006) . The Bezier method, as implemented here, requires four control points (2-D coordinates) as input, along with their corresponding times; these four points are passed to the function as a dataframe with four rows, and three columns corresponding to x, y, & t (the names of the columns do not matter). The function interpolates the movement trajectory for the times indicated by parameter \code{t.slice}. The \code{t.slice} times must lie between the time of the 2nd and 3rd control point. Times can be passed in as either POSIX-class objects or numeric, but not a mixture of both.
}
\examples{
data(contrived)
xyt <- contrived
###times for interpolation
t.slice <- c(1.5,2,2.5,3,3.5,4,4.5,5,5.5)
a <- bezier(xyt,t.slice)
plot(xyt[,1],xyt[,2],pch=20)
points(a[,1],a[,2])

}
\references{
Long, JA (2015) Kinematic interpolation of movement data. \emph{International Journal of Geographical Information Science}. DOI: 10.1080/13658816.2015.1081909. \cr \cr
Tremblay, YC \emph{et al.} (2006) Interpolation of animal tracking data in a fluid environment. \emph{Journal of Experimental Biology}. 209(1): 128-140.
}
\keyword{interpolation}

