% Generated by roxygen2 (4.0.1): do not edit by hand
\name{kinematic}
\alias{kinematic}
\title{Interpolate using kinematic interpolation}
\usage{
kinematic(xytvv, t.slice)
}
\arguments{
\item{xytvv}{a 2x5 dataframe containing the coordinates, times, and initial and final velocities (as 2D vectors) of the two points to be interpolated between, often termed the anchor points. Each row of the dataframe should be arranged as x, y, t, vx, vy.}

\item{t.slice}{a single time (POSIX or numeric), or list of times, to be interpolated for. The times must lie between those of the points in \code{xytv}.}
}
\value{
The function returns a dataframe (with \code{nrow = length{t.slice}}) corresponding to the interpolated locations.
}
\description{
Perform kinematic path interpolation on a movement dataset. Kinematic path interpolation was introduced in the paper Long (2015). Kinematic interpolation is appropriate for fast moving objects, recorded with relatively high resolution tracking data.
}
\details{
Kinematic interpolation requires the user to input the coordinates of the anchor points between which the interpolation is occurring, as well as initial and final velocities associate with the anchor points. In practice, these velocities may be explicitly known, or estimated from the tracking data.
}
\examples{
data(contrived)
xyt <- contrived
###times for interpolation
t.slice <- c(1.5,2,2.5,3,3.5,4,4.5,5,5.5)
### add velocities for kinematic analysis
xytvv <- cbind(xyt[2:3,],rbind(c(0,3),c(3,0)))
a <- kinematic(xytvv,t.slice)
plot(xyt[,1],xyt[,2],pch=20)
points(a[,1],a[,2])
}
\references{
Long, JA (2015) Kinematic interpolation of movement data. \emph{International Journal of Geographical Information Science}. DOI: 10.1080/13658816.2015.1081909.
}
\keyword{interpolation}

