% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{linear}
\alias{linear}
\title{Interpolate using linear interpolation}
\usage{
linear(xyt, t.slice)
}
\arguments{
\item{xyt}{a 2x3 dataframe containing the coordinates and times of the two points to be interpolated between, often termed the anchor points. Each row of the dataframe should be arranged as x, y, t.}

\item{t.slice}{a single time (POSIX or numeric), or list of times, to be interpolated for. The times must lie between those of the points in \code{xyt}.}
}
\value{
The function returns a dataframe (with \code{nrow = length{t.slice}}) corresponding to the interpolated locations.
}
\description{
Perform linear path interpolation on a movement dataset.
}
\details{
Linear interpolation is the simplest, and most commonly employed methods for path interpolation.
}
\examples{
data(contrived)
xyt <- contrived
###times for interpolation
t.slice <- c(1.5,2,2.5,3,3.5,4,4.5,5,5.5)
a <- linear(xyt[2:3,],t.slice)
plot(xyt[,1],xyt[,2],pch=20)
points(a[,1],a[,2])

}
\keyword{interpolation}

