% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathInterpolatR-package.R
\docType{package}
\name{PathInterpolatR-package}
\alias{PathInterpolatR-package}
\title{PathInterpolatR: Methods for Path Interpolation}
\description{
The package \code{PathInterpolatR} provides tools for performing path interpolation on movement data (such as GPS tracking data). These tools were developed in support of the paper Long (2015) and extend the work of other authors (see references within).
}
\details{
Currently the code has been developed for the simple scenario of interpolation between two points (where points on either side may or may not be known). Eventually, I will extend the code to facilitate straightforward implementation of interpolation with common movement data structures from other R packages (e.g., \code{adehabitat} and \code{MOVE}).
}
\author{
Jed Long
}
\references{
Long, JA (2015) Kinematic interpolation of movement data. \emph{International Journal of Geographical Information Science}. DOI: 10.1080/13658816.2015.1081909.
}

